requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20121119
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.5
django-allauth==0.33.0
django-cors-headers==2.1.0
django-import-export==0.5.1
django-jet==1.0.6
django-seed==0.1.7
djangorestframework==3.6.4
dnspython==1.15.0
email-validator==1.0.2
et-xmlfile==1.0.1
Faker==0.8.3
idna==2.6
jdcal==1.3
mysqlclient==1.3.12
oauthlib==2.0.4
odfpy==1.3.5
openpyxl==2.4.8
psycopg2==2.7.3.1
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.2
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
static3==0.7.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
virtualenv==15.1.0
whitenoise==3.3.1
xlrd==1.1.0
xlwt==1.3.0
